post section were done

+++ app/Http/Controllers/AdminUserController.php
@@ -130,7 +130,11 @@ class AdminUserController extends Controller
     {
         //
         $user=User::findOrFail($id);
-        unlink(public_path().$user->photo->file);
+        if($user->photo)
+        {
+            unlink(public_path().$user->photo->file);
+        }
+       
         $user->delete();
         Session::flash('deleted_user','User has been delted');
         return redirect("/admin/users");
